Java seperated out, commented and obsolete code/classes removed
Added comments to everything that needs them, created a new java class
which is called 'TableFiller.java' which resets the table and adds all
rows and columns MyUserInterface.java. also the control now only calls
the interface ot be created. While myUserInterface deals with the new
data submitted through GUI and DatabaseReader which deals with queries